fix(cli): clear stale proposal-record output and refuse aliased paths
