+++ crates/adoc-cli/src/commands/semantic_executor.rs
@@ -1,6 +1,7 @@
 use std::fs;
 use std::path::{Path, PathBuf};
 
+use crate::commands::artifact_paths::{ensure_distinct_paths, remove_stale};
 use adoc_core::{
     DiagnosticCode, HumanReviewExpectedBindings, SemanticExecutorError,
     build_semantic_context_from_document, complete_semantic_execution, fail_semantic_execution,
@@ -190,47 +191,6 @@ fn record_failure(
     eprintln!("error[{code}] {message}");
     print!("{receipt}");
     2
-}
-
-fn remove_stale(path: &Path) -> Result<(), String> {
-    match fs::remove_file(path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(format!(
-            "could not remove stale output {}: {error}",
-            path.display()
-        )),
-    }
-}
-
-fn ensure_distinct_paths(paths: &[&Path]) -> Result<(), String> {
-    let identities = paths
-        .iter()
-        .map(|path| path_identity(path))
-        .collect::<Vec<_>>();
-    if identities
-        .iter()
-        .enumerate()
-        .any(|(index, path)| identities[..index].contains(path))
-    {
-        return Err("semantic input and output artifact paths must be distinct".to_string());
-    }
-    Ok(())
-}
-
-fn path_identity(path: &Path) -> PathBuf {
-    fs::canonicalize(path).unwrap_or_else(|_| {
-        let parent = path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-            .unwrap_or_else(|| Path::new("."));
-        fs::canonicalize(parent)
-            .map(|parent| {
-                path.file_name()
-                    .map_or(parent.clone(), |name| parent.join(name))
-            })
-            .unwrap_or_else(|_| path.to_path_buf())
-    })
 }
 
 fn fail(message: &str) -> i32 {
